requirements.txt
# Core Frameworks
streamlit==1.32.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1

# Vector Database & Storage
chromadb==0.4.24
langchain-chroma==0.1.1
pydantic==2.7.1

# AI & Embeddings
sentence-transformers==2.7.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1

# Fix for Protobuf/Linux issues
protobuf==3.20.3
urllib3==2.2.1
packaging==23.2

pypdf==6.10.2